requirements.txt
Django==5.1.1
asgiref==3.8.1
sqlparse==0.5.1
dj-database-url==3.0.1
psycopg2-binary==2.9.10
whitenoise==6.7.0

djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0

google-auth==2.43.0
requests==2.32.5

gunicorn==23.0.0
PyJWT==2.9.0
python-dotenv==1.1.1
